internal/ui: implement page updates via an interface
