+++ internal/ui/trayscale.cmb
@@ -48,6 +48,7 @@
 	(2,126,"GtkLabel","TxBytes",125,None,None,None,None,None),
 	(2,135,"GtkDropTarget","DropTarget",None,None,None,None,None,None),
 	(2,136,"GtkButton","SendFileButton",83,None,None,None,None,None),
+	(2,137,"AdwPreferencesGroup","AdvertisedRoutesGroup",74,None,None,None,3,None),
 	(3,1,"AdwPreferencesWindow","PreferencesWindow",None,None,None,None,None,None),
 	(3,2,"AdwPreferencesPage",None,1,None,None,None,None,None),
 	(3,3,"AdwPreferencesGroup",None,2,None,None,None,None,None),
@@ -149,6 +150,7 @@
 	(2,136,"GtkButton","icon-name","document-send-symbolic",None,None,None,None,None,None,None,None,None),
 	(2,136,"GtkWidget","margin-bottom","12",None,None,None,None,None,None,None,None,None),
 	(2,136,"GtkWidget","margin-top","12",None,None,None,None,None,None,None,None,None),
+	(2,137,"AdwPreferencesGroup","title","Advertised Routes",None,None,None,None,None,None,None,None,None),
 	(3,3,"AdwPreferencesGroup","title","General",None,None,None,None,None,None,None,None,None),
 	(3,4,"AdwActionRow","activatable-widget","12",None,None,None,None,None,None,None,None,None),
 	(3,4,"AdwActionRow","subtitle","If enabled, an icon will be added to the system tray",None,None,None,None,None,None,None,None,None),
